refactor(search): the branch table moves beside the gates it obeys
store.go crossed the 500-line ceiling when the catalog branches landed.
It splits at the seam that was already there: what a searchable entity
IS, and who may see one, is a decision read three ways — the lexical
union, the vector lane and the plan compiler each build their own SQL
from these rows — while store.go is one of those three readers.

The context-anchor gate AST-parses the branch table, so it follows the
table to its new file rather than keeping a path that would now parse a
file with no branches in it and hold the contract against an empty set.
